pyproject.toml
[tool.poetry]
name = "giaco-dev-backend"
version = "0.1.0"
description = "portofolio web dev of giacomo tolari"
authors = ["Giacomo Tolari <giacomotolari@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
black = "^22.12.0"
django = "^5.1.2"
django-redis = "^5.4.0"
redis = "^5.1.1"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"
django-multiselectfield = "^0.1.13"

[tool.black]
line-length = 105
target-version = ['py39']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
